chatbot/requirements.txt
python-dotenv==1.0.1
openai>=1.37.0
pinecone>=5.0.0
pypdf==4.0.1
uvicorn>=0.27.0
fastapi>=0.110.0
python-multipart==0.0.9
